Fixed a crash after clicking settings on MacBook without a notch
Fixed a crash in NotchSettingsViewController that occurred on Macs without a notch by ensuring NSImage(systemSymbolName:) only initializes when the symbol exists.

Somewhat related:
I had to replace the "import SwiftUICore" imports in three files with "import SwiftUI" since SwiftUICore is not included in the repo. Feel free to test that on new devices as well.